refactored into 2 files
